requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
-e git+https://github.com/ajacobsen/django-video-encoding.git@f88b4b296024a2f1dfaadfcfcc3744f36ef5812d#egg=django_video_encoding  # via -r requirements.in
asgiref==3.2.3            # via django
click==7.0                # via rq
django-appconf==1.0.3     # via -r requirements.in
django-rq==2.3.0          # via -r requirements.in
django==3.0.3             # via -r requirements.in, django-appconf, django-rq
fixedint==0.1.5           # via -r requirements.in
pillow==7.0.0             # via -r requirements.in
psycopg2-binary==2.8.4    # via -r requirements.in
python-dotenv==0.12.0     # via -r requirements.in
python-magic==0.4.15      # via -r requirements.in
pytz==2019.3              # via django
redis==3.4.1              # via django-rq, rq
rq==1.2.2                 # via django-rq
shutilwhich==1.1.0        # via -r requirements.in
six==1.14.0               # via django-appconf
sqlparse==0.3.0           # via django
wheel==0.34.2             # via -r requirements.in
